services/content/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
alembic==1.12.1
python-dotenv==1.0.0
httpx==0.25.2
pika==1.3.2
markdown==3.5.1
python-slugify==8.0.1
setuptools>=65.5.1 